+++ Package.swift
@@ -42,11 +42,7 @@ let package = Package(
             ],
             exclude: [
                 "build.sh",
-                "swiftgen.yml",
                 "Data/GQL/Schema",
-            ],
-            resources: [
-                .process("Resources"),
             ]
         ),
         .target(