+++ Sources/NablaMessagingCore/build.sh
@@ -25,11 +25,3 @@ run_cached apollo-codegen "${input_output_files[*]}" \
   /usr/bin/xcrun --sdk macosx swift run ApolloCodegen generate
 
 cd "$DIRNAME"
-
-# SwiftGen
-GENERATED_FOLDER="$DIRNAME"/Generated
-echo "Create Generated folder in $GENERATED_FOLDER if not existing"
-mkdir -p "$GENERATED_FOLDER"
-
-echo "Execute swiftgen"
-mint run -m ../../Mintfile swiftgen config run --config "$DIRNAME"/swiftgen.yml